+++ src/ApiObjects/Collections/UnknownCollection.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace Evgeek\Moysklad\ApiObjects\Collections;
 
 use Evgeek\Moysklad\ApiObjects\AbstractUnknownApiObject;
-use Evgeek\Moysklad\ApiObjects\Collections\Traits\SetMetaTrait;
+use Evgeek\Moysklad\ApiObjects\Collections\Traits\SetMetaCollectionTrait;
 
 class UnknownCollection extends AbstractUnknownApiObject
 {
-    use SetMetaTrait;
+    use SetMetaCollectionTrait;
 }